Fix MySQL mappers parameter types

+++ source/src/main/java/fci/swe2/onlineshopapi/Account.java
@@ -7,7 +7,7 @@ abstract class Account {
     protected String password;
     protected RegistrationBehaviour registrationBehaviour;
 
-    protected int getUserID(){
+    protected long getUserID(){
         return this.userID;
     }
